# Conflicts:
#	main-service/src/main/java/ru/practicum/ewm/event/EventRepository.java
#	main-service/src/main/java/ru/practicum/ewm/event/EventServiceImpl.java

+++ main-service/src/main/java/ru/practicum/ewm/event/UpdateEventAdminRequest.java
@@ -29,7 +29,7 @@ record UpdateEventAdminRequest(
         String description,
 
         @PositiveOrZero
-        Integer participantLimit,
+        Long participantLimit,
 
         Boolean paid,
         Boolean requestModeration,
